combined navigation test
